moved bin from armus/bin to root of project
